move Py4J server to integration package
