+++ src/main/java/denoptim/main/Main.java
@@ -37,9 +37,9 @@ import denoptim.exception.ExceptionUtils;
 import denoptim.files.FileFormat;
 import denoptim.files.FileUtils;
 import denoptim.gui.GUI;
+import denoptim.integration.python.Py4JGetawayServer;
 import denoptim.logging.Version;
 import denoptim.programs.combinatorial.FragSpaceExplorer;
-import denoptim.programs.deamon.Py4JGetawayServer;
 import denoptim.programs.denovo.GARunner;
 import denoptim.programs.fitnessevaluator.FitnessRunner;
 import denoptim.programs.fragmenter.Fragmenter;
